+++ apps/main/src/dex/components/PageCreatePool/index.tsx
@@ -21,11 +21,11 @@ import { useNetworks } from '@/dex/entities/networks'
 import { curveProps } from '@/dex/lib/utils'
 import { useStore } from '@/dex/store/useStore'
 import { CurveApi, ChainId } from '@/dex/types/main.types'
-import { useLayoutStore } from '@evm-ui/features/layout/store'
 import { Box } from '@legacy-ui/Box'
 import { Button } from '@legacy-ui/Button'
 import { Icon } from '@legacy-ui/Icon'
 import { Spinner } from '@legacy-ui/Spinner'
+import { useLayoutStore } from '@ui/features/layout/layout/store'
 import { t } from '@ui/lib/i18n'
 
 type Props = {
