+++ apps/main/src/dex/components/PageDashboard/components/Summary.tsx
@@ -9,13 +9,13 @@ import { TotalRecurrence as SummaryRecurrence } from '@/dex/components/PageDashb
 import { SummaryTotal } from '@/dex/components/PageDashboard/components/SummaryTotal'
 import { useDashboardContext } from '@/dex/components/PageDashboard/dashboardContext'
 import { useStore } from '@/dex/store/useStore'
-import { useLayoutStore } from '@evm-ui/features/layout'
 import { Tabs } from '@evm-ui/shared/ui/Tabs/Tabs'
 import { Box } from '@legacy-ui/Box'
 import { SpinnerWrapper } from '@legacy-ui/Spinner'
 import { Stats } from '@legacy-ui/Stats'
 import { shortenAccount } from '@legacy-ui/utils'
 import { breakpoints } from '@legacy-ui/utils/responsive'
+import { useLayoutStore } from '@ui/features/layout/layout'
 import { t } from '@ui/lib/i18n'
 
 const menu = [
