+++ apps/main/src/dex/components/PageRouterSwap/index.tsx
@@ -23,7 +23,6 @@ import { toTokenOption } from '@/dex/utils'
 import { getSlippageImpact } from '@/dex/utils/utilsSwap'
 import type { Chain } from '@curvefi/prices-api'
 import { notify } from '@evm-ui/features/connect-wallet'
-import { useLayoutStore } from '@evm-ui/features/layout'
 import { TokenList, TokenSelector, useTokenSelectorData } from '@evm-ui/features/select-token'
 import { useUserProfileStore } from '@evm-ui/features/user-profile'
 import { usePageVisibleInterval } from '@evm-ui/hooks/usePageVisibleInterval'
@@ -49,6 +48,7 @@ import type { Address } from '@primitives/address.utils'
 import type { Decimal } from '@primitives/decimal.utils'
 import { assert, maybe, maybes, notFalsy } from '@primitives/objects.utils'
 import type { RouterRouteResponse } from '@primitives/router.utils'
+import { useLayoutStore } from '@ui/features/layout/layout'
 import { mapQuery, q, toQuery } from '@ui/features/queries/util'
 import { SizesAndSpaces } from '@ui/features/themes/design/1_sizes_spaces'
 import { useSwitch } from '@ui/hooks/useSwitch'
