+++ apps/main/src/llamalend/features/market-list/MarketExpandedPanel.tsx
@@ -2,7 +2,6 @@ import { type FunctionComponent, ReactNode, useMemo } from 'react'
 import { NET_SUPPLY_RATE_TITLE } from '@/llamalend/constants'
 import { tokenMetric } from '@/llamalend/llama.utils'
 import type { LlamaMarketRow } from '@/llamalend/queries/market-list/llama-market-stats'
-import { useLayoutStore } from '@evm-ui/features/layout'
 import { type ExpandedPanelComponent } from '@evm-ui/shared/ui/DataTable/ExpansionRow'
 import { Metric } from '@evm-ui/shared/ui/Metric'
 import { MarketRateType } from '@evm-ui/types/market'
@@ -12,6 +11,7 @@ import CardHeader, { CardHeaderProps } from '@mui/material/CardHeader'
 import Grid from '@mui/material/Grid'
 import Stack from '@mui/material/Stack'
 import Typography from '@mui/material/Typography'
+import { useLayoutStore } from '@ui/features/layout/layout'
 import { constQ } from '@ui/features/queries/util'
 import { SizesAndSpaces } from '@ui/features/themes/design/1_sizes_spaces'
 import { useIsTiny } from '@ui/hooks/useBreakpoints'
