+++ packages/evm-ui/src/hooks/useActiveSection.ts
@@ -1,8 +1,8 @@
 import { useEffect, useRef } from 'react'
-import { useLayoutStore } from '@evm-ui/features/layout'
 import { useLocation, useNavigate } from '@evm-ui/hooks/router'
 import { notFalsy } from '@primitives/objects.utils'
 import { useRouterState } from '@tanstack/react-router'
+import { useLayoutStore } from '@ui/features/layout/layout'
 import { useResizeObserver } from '@ui/hooks/useResizeObserver'
 
 /** Height in pixels of the viewport band used to determine the active section. */
