+++ packages/evm-ui/src/hooks/useIncreasingLength.ts
@@ -1,5 +1,5 @@
 import { useEffect, useState } from 'react'
-import { setTimeoutInterval } from '@evm-ui/utils/timers'
+import { setTimeoutInterval } from '@ui/lib/timers'
 
 type IncreasingLengthOptions = {
   initialLength?: number
