+++ packages/evm-ui/src/pages/ErrorPage.tsx
@@ -1,5 +1,4 @@
 import { type ElementType, useCallback, useState } from 'react'
-import { useLayoutStore } from '@evm-ui/features/layout'
 import { ErrorReportModal } from '@evm-ui/features/report-error'
 import { getBoundaryErrorSubtitle } from '@evm-ui/utils/errors'
 import Button from '@mui/material/Button'
@@ -7,6 +6,7 @@ import Stack from '@mui/material/Stack'
 import Typography from '@mui/material/Typography'
 import type { Address } from '@primitives/address.utils'
 import { RouterLink } from '@ui/components/RouterLink'
+import { useLayoutStore } from '@ui/features/layout/layout'
 import { persister, queryClient } from '@ui/features/queries/query-client'
 import { SizesAndSpaces } from '@ui/features/themes/design/1_sizes_spaces'
 import { useSwitch } from '@ui/hooks/useSwitch'
