+++ packages/evm-ui/src/shared/ui/DataTable/LegacyDataTable.tsx
@@ -1,6 +1,5 @@
 /// <reference types="./DataTable.d.ts" />
 import { type ReactNode, useEffect, useEffectEvent, useMemo, useRef } from 'react'
-import { useLayoutStore } from '@evm-ui/features/layout'
 import { TablePagination } from '@evm-ui/shared/ui/DataTable/TablePagination'
 import Box from '@mui/material/Box'
 import { Theme } from '@mui/material/styles'
@@ -12,6 +11,7 @@ import TableHead from '@mui/material/TableHead'
 import TableRow from '@mui/material/TableRow'
 import type { ReactTable, RowData } from '@tanstack/react-table'
 import { WithWrapper } from '@ui/components/WithWrapper'
+import { useLayoutStore } from '@ui/features/layout/layout'
 import { SizesAndSpaces } from '@ui/features/themes/design/1_sizes_spaces'
 import { useIsMobile } from '@ui/hooks/useBreakpoints'
 import { t } from '@ui/lib/i18n'
