+++ packages/evm-ui/src/widgets/DetailPageLayout/DetailPageLayout.tsx
@@ -1,9 +1,9 @@
 import type { ReactNode } from 'react'
-import { useLayoutStore } from '@evm-ui/features/layout'
 import Box from '@mui/material/Box'
 import Grid from '@mui/material/Grid'
 import Stack, { StackProps } from '@mui/material/Stack'
 import { WithWrapper } from '@ui/components/WithWrapper'
+import { useLayoutStore } from '@ui/features/layout/layout'
 import { mapBreakpoints } from '@ui/features/themes/basic-theme/basic-theme'
 import { SizesAndSpaces } from '@ui/features/themes/design/1_sizes_spaces'
 import { useIsMobile } from '@ui/hooks/useBreakpoints'
