+++ packages/evm-ui/src/widgets/Header/MobileHeader.tsx
@@ -1,5 +1,4 @@
 import { useEffect, useMemo } from 'react'
-import { useLayoutStore } from '@evm-ui/features/layout'
 import { usePathname } from '@evm-ui/hooks/router'
 import { APP_LINK, routeToPage } from '@evm-ui/shared/routes'
 import { GlobalBanner } from '@evm-ui/shared/ui/GlobalBanner'
@@ -8,6 +7,7 @@ import Drawer from '@mui/material/Drawer'
 import Stack from '@mui/material/Stack'
 import Toolbar from '@mui/material/Toolbar'
 import { recordEntries } from '@primitives/objects.utils'
+import { useLayoutStore } from '@ui/features/layout/layout'
 import { MOBILE_SIDEBAR_WIDTH } from '@ui/features/themes/components'
 import { useSwitch } from '@ui/hooks/useSwitch'
 import { t } from '@ui/lib/i18n'
