+++ packages/evm-ui/src/widgets/Toast/Toast.tsx
@@ -1,9 +1,9 @@
 import { useEffect, useState } from 'react'
-import { useLayoutStore } from '@evm-ui/features/layout'
 import Alert from '@mui/material/Alert'
 import AlertTitle from '@mui/material/AlertTitle'
 import Snackbar from '@mui/material/Snackbar'
 import Stack from '@mui/material/Stack'
+import { useLayoutStore } from '@ui/features/layout/layout'
 import { Duration } from '@ui/features/themes/design/0_primitives'
 import { SizesAndSpaces } from '@ui/features/themes/design/1_sizes_spaces'
 import { watchToasts, type ToastItem } from './toast.util'
